Prefix target names with 'GraphViz' to avoid collisions
