+++ Sources/GraphVizBuilder/Extensions/Edge+Extensions.swift
@@ -1,6 +1,6 @@
-import Core
+import GraphVizCore
 
-extension Node {
+extension Edge {
     public subscript<T>(dynamicMember member: WritableKeyPath<Attributes, T>) -> (T) -> Self {
         get {
             var mutableSelf = self
