+++ Sources/GraphVizBuilder/Extensions/Node+Extensions.swift
@@ -1,6 +1,6 @@
-import Core
+import GraphVizCore
 
-extension Edge {
+extension Node {
     public subscript<T>(dynamicMember member: WritableKeyPath<Attributes, T>) -> (T) -> Self {
         get {
             var mutableSelf = self
