+++ Sources/GraphVizCore/Edge.swift
