+++ Sources/GraphVizCore/Node.swift
