+++ Sources/GraphVizCore/Supporting Types/AspectRatio.swift
