+++ Sources/GraphVizCore/Supporting Types/Color.swift
