+++ Sources/GraphVizCore/Supporting Types/Ordering.swift
