+++ Sources/GraphVizCore/Supporting Types/Rectangle.swift
