+++ Sources/GraphVizCore/Supporting Types/Size.swift
