+++ Sources/GraphVizTools/Error.swift
