+++ Sources/GraphVizTools/Extensions/Graph+Rendering.swift
@@ -1,7 +1,7 @@
 import Foundation
 import Dispatch
 
-import Core
+import GraphVizCore
 
 extension Graph {
     /**
