+++ Tests/GraphVizTests/AttributesTest.swift
@@ -1,5 +1,5 @@
 import XCTest
-@testable import Core
+@testable import GraphVizCore
 
 final class AttributesTests: XCTestCase {
     func testGraphAttributesCompleteness() {
